# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "ok",
                    "db": "up"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "summary": "Create an account and get tokens.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "summary": "Log in with email and password.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token expired, reused or revoked"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "summary": "Exchange a refresh token for a new token pair (the old refresh token stops working).",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current session's refresh token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/users/me/events": {
      "get": {
        "operationId": "UsersController_hosting",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyEventsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/users/me/rsvps": {
      "get": {
        "operationId": "UsersController_rsvps",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyEventsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/events": {
      "get": {
        "operationId": "EventsController_list",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Only events starting at/after this time. Defaults to now.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Only events starting before this time.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search on title and location.",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "creatorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page's `nextCursor`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          }
        },
        "summary": "List upcoming published events (cursor-paginated). Public.",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an event. You become its host.",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one event. Send a token to also get `myRsvpStatus`.",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not the host"
          },
          "409": {
            "description": "Stale version, or capacity below attendees"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit an event. Host only.",
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "Not the host"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel (soft-delete) an event. Host only.",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}/rsvp": {
      "post": {
        "operationId": "RsvpsController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RsvpResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Event cancelled or already started"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "RSVP to an event. Returns `going` if you got a seat, otherwise `waitlisted`.\nSafe to retry — calling it twice never double-books.",
        "tags": [
          "rsvps"
        ]
      },
      "delete": {
        "operationId": "RsvpsController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RsvpResponse"
                }
              }
            }
          },
          "404": {
            "description": "You haven't RSVP'd"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel your RSVP. Your seat goes to the next person on the waitlist.",
        "tags": [
          "rsvps"
        ]
      }
    },
    "/api/v1/events/{id}/attendees": {
      "get": {
        "operationId": "RsvpsController_attendees",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 30,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendeePage"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "People going to this event, in the order they got their seat. Public.",
        "tags": [
          "rsvps"
        ]
      }
    }
  },
  "info": {
    "title": "Events API",
    "description": "Events + RSVP API. NestJS · PostgreSQL · JWT.\n\n**Try it:** `POST /api/v1/auth/login` with `demo@events.dev` / `Password123!`,\ncopy `accessToken`, click **Authorize**, then call any endpoint.\n\nErrors always look like `{ statusCode, error, message, path, timestamp }`.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "you@example.com",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "At least 8 characters.",
            "example": "Password123!",
            "minLength": 8,
            "maxLength": 128
          },
          "name": {
            "type": "string",
            "example": "Asha Rao",
            "minLength": 1,
            "maxLength": 80
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "AuthTokensResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number",
            "description": "Access-token lifetime in seconds"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUser"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "demo@events.dev",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "hostingCount": {
            "type": "number",
            "description": "Events you host (not cancelled)"
          },
          "goingCount": {
            "type": "number",
            "description": "Events you're going to"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt",
          "hostingCount",
          "goingCount"
        ]
      },
      "HostResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "EventResponse": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "draft",
              "published",
              "cancelled"
            ],
            "type": "string"
          },
          "myRsvpStatus": {
            "nullable": true,
            "description": "The caller's RSVP status — present only when a valid token is sent.",
            "enum": [
              "going",
              "waitlisted",
              "cancelled"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "nullable": true
          },
          "goingCount": {
            "type": "number"
          },
          "seatsLeft": {
            "type": "number",
            "nullable": true,
            "description": "null when capacity is unlimited"
          },
          "version": {
            "type": "number"
          },
          "host": {
            "$ref": "#/components/schemas/HostResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "title",
          "description",
          "location",
          "startsAt",
          "endsAt",
          "capacity",
          "goingCount",
          "seatsLeft",
          "version",
          "host",
          "createdAt",
          "updatedAt"
        ]
      },
      "MyEventsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "EventPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pass as `cursor` to get the next page; null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "`draft` events are visible only to their host.",
            "enum": [
              "draft",
              "published"
            ],
            "type": "string",
            "default": "published"
          },
          "title": {
            "type": "string",
            "example": "Bangalore Backend Meetup",
            "minLength": 3,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Talks on Postgres internals and NestJS at scale.",
            "maxLength": 5000
          },
          "location": {
            "type": "string",
            "example": "91springboard, Koramangala",
            "maxLength": 200
          },
          "startsAt": {
            "type": "string",
            "description": "ISO 8601, must be in the future.",
            "example": "2026-10-10T12:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "description": "ISO 8601, after startsAt.",
            "example": "2026-10-10T15:30:00.000Z"
          },
          "capacity": {
            "type": "number",
            "nullable": true,
            "description": "Max attendees. Omit or null for unlimited.",
            "example": 50,
            "minimum": 1,
            "maximum": 100000
          }
        },
        "required": [
          "title",
          "startsAt",
          "endsAt"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "`draft` events are visible only to their host.",
            "enum": [
              "draft",
              "published"
            ],
            "type": "string",
            "default": "published"
          },
          "title": {
            "type": "string",
            "example": "Bangalore Backend Meetup",
            "minLength": 3,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Talks on Postgres internals and NestJS at scale.",
            "maxLength": 5000
          },
          "location": {
            "type": "string",
            "example": "91springboard, Koramangala",
            "maxLength": 200
          },
          "startsAt": {
            "type": "string",
            "description": "ISO 8601, must be in the future.",
            "example": "2026-10-10T12:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "description": "ISO 8601, after startsAt.",
            "example": "2026-10-10T15:30:00.000Z"
          },
          "capacity": {
            "type": "number",
            "nullable": true,
            "description": "Max attendees. Omit or null for unlimited.",
            "example": 50,
            "minimum": 1,
            "maximum": 100000
          },
          "version": {
            "type": "number",
            "description": "Optimistic lock: send the `version` you last read. If someone else edited\nthe event since, you get 409 instead of silently overwriting their change."
          }
        }
      },
      "RsvpResponse": {
        "type": "object",
        "properties": {
          "status": {
            "description": "`going` if you got a seat, `waitlisted` if the event is full.",
            "enum": [
              "going",
              "waitlisted",
              "cancelled"
            ],
            "type": "string"
          },
          "eventId": {
            "type": "string"
          },
          "goingCount": {
            "type": "number"
          },
          "seatsLeft": {
            "type": "number",
            "nullable": true,
            "description": "null when capacity is unlimited"
          }
        },
        "required": [
          "status",
          "eventId",
          "goingCount",
          "seatsLeft"
        ]
      },
      "Attendee": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When this person got their seat"
          }
        },
        "required": [
          "userId",
          "name",
          "joinedAt"
        ]
      },
      "AttendeePage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendee"
            }
          },
          "total": {
            "type": "number"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "total",
          "nextCursor"
        ]
      }
    }
  }
}